ralaciomaneto codigo matricula adicionado

+++ src/main/java/br/edu/ufersa/sys_scholar/api/dto/AbstractAlunoDTO.java
@@ -9,7 +9,7 @@ public abstract class AbstractAlunoDTO {
 
     protected Long id;
 
-    protected Integer codigo;
+    protected Long codigo;
 
     protected String nome;
 }
